Move Bin and SymInfo into their own files
